+++ packages/auth-construct/src/construct.test.ts
@@ -11,6 +11,7 @@ import {
 import {
   CfnIdentityPool,
   CfnUserPoolClient,
+  ProviderAttribute,
   UserPool,
   UserPoolClient,
 } from 'aws-cdk-lib/aws-cognito';
@@ -31,7 +32,11 @@ const facebookClientSecret = 'facebookClientSecret';
 const oidcClientId = 'oidcClientId';
 const oidcClientSecret = 'oidcClientSecret';
 const oidcIssuerUrl = 'https://mysampleoidcissuer.com';
-const oidcProviderName = 'myOidcProvider';
+const oidcProviderName = 'MyOidcProvider';
+const oidcClientId2 = 'oidcClientId2';
+const oidcClientSecret2 = 'oidcClientSecret2';
+const oidcIssuerUrl2 = 'https://mysampleoidcissuer2.com';
+const oidcProviderName2 = 'MyOidcProvider2';
 const ExpectedGoogleIDPProperties = {
   ProviderDetails: {
     authorize_scopes: 'profile',
@@ -79,6 +84,17 @@ const ExpectedOidcIDPProperties = {
     oidc_issuer: oidcIssuerUrl,
   },
   ProviderName: oidcProviderName,
+  ProviderType: 'OIDC',
+};
+const ExpectedOidcIDPProperties2 = {
+  ProviderDetails: {
+    attributes_request_method: 'GET',
+    authorize_scopes: 'openid',
+    client_id: oidcClientId2,
+    client_secret: oidcClientSecret2,
+    oidc_issuer: oidcIssuerUrl2,
+  },
+  ProviderName: oidcProviderName2,
   ProviderType: 'OIDC',
 };
 const samlProviderName = 'samlProviderName';
@@ -1230,16 +1246,39 @@ void describe('Auth construct', () => {
     void it('supports oidc and email', () => {
       const app = new App();
       const stack = new Stack(app);
+      const authorizationURL = 'http://localhost:3000/authorization';
+      const jwksURI = 'https://localhost:3000/jwksuri';
+      const tokensURL = 'http://localhost:3000/token';
+      const userInfoURL = 'http://localhost:3000/userinfo';
+      const mockIdentifiers = ['one', 'two'];
+      const mockScopes = ['scope1', 'scope2'];
+      const attributeRequestMethod = 'POST';
       new AmplifyAuth(stack, 'test', {
         loginWith: {
           email: true,
           externalProviders: {
-            oidc: {
-              clientId: oidcClientId,
-              clientSecret: oidcClientSecret,
-              issuerUrl: oidcIssuerUrl,
-              name: oidcProviderName,
-            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+                attributeMapping: {
+                  email: {
+                    attributeName: 'email',
+                  },
+                },
+                attributeRequestMethod: attributeRequestMethod,
+                endpoints: {
+                  authorization: authorizationURL,
+                  jwksUri: jwksURI,
+                  token: tokensURL,
+                  userInfo: userInfoURL,
+                },
+                identifiers: mockIdentifiers,
+                scopes: mockScopes,
+              },
+            ],
             callbackUrls: ['https://redirect.com'],
             logoutUrls: ['https://logout.com'],
           },
@@ -1250,10 +1289,25 @@ void describe('Auth construct', () => {
         UsernameAttributes: ['email'],
         AutoVerifiedAttributes: ['email'],
       });
-      template.hasResourceProperties(
-        'AWS::Cognito::UserPoolIdentityProvider',
-        ExpectedOidcIDPProperties
-      );
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        AttributeMapping: {
+          email: 'email',
+        },
+        IdpIdentifiers: mockIdentifiers,
+        ProviderDetails: {
+          attributes_request_method: attributeRequestMethod,
+          attributes_url: userInfoURL,
+          authorize_scopes: mockScopes.join(' '),
+          authorize_url: authorizationURL,
+          client_id: oidcClientId,
+          client_secret: oidcClientSecret,
+          jwks_uri: jwksURI,
+          oidc_issuer: oidcIssuerUrl,
+          token_url: tokensURL,
+        },
+        ProviderName: oidcProviderName,
+        ProviderType: 'OIDC',
+      });
       template.hasResourceProperties('AWS::Cognito::IdentityPool', {
         OpenIdConnectProviderARNs: [
           Match.objectEquals({
@@ -1267,7 +1321,90 @@ void describe('Auth construct', () => {
                 ':oidc-provider/cognito-idp.',
                 { Ref: 'AWS::Region' },
                 '.amazonaws.com/',
-                { Ref: 'testOidcIDP12B3582F' },
+                { Ref: 'testMyOidcProviderOidcIDP837BDEAD' },
+              ],
+            ],
+          }),
+        ],
+      });
+    });
+    void it('oidc defaults to GET for oidc method', () => {
+      const app = new App();
+      const stack = new Stack(app);
+      const authorizationURL = 'http://localhost:3000/authorization';
+      const jwksURI = 'https://localhost:3000/jwksuri';
+      const tokensURL = 'http://localhost:3000/token';
+      const userInfoURL = 'http://localhost:3000/userinfo';
+      const mockIdentifiers = ['one', 'two'];
+      const mockScopes = ['scope1', 'scope2'];
+      new AmplifyAuth(stack, 'test', {
+        loginWith: {
+          email: true,
+          externalProviders: {
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+                attributeMapping: {
+                  email: {
+                    attributeName: 'email',
+                  },
+                },
+                endpoints: {
+                  authorization: authorizationURL,
+                  jwksUri: jwksURI,
+                  token: tokensURL,
+                  userInfo: userInfoURL,
+                },
+                identifiers: mockIdentifiers,
+                scopes: mockScopes,
+              },
+            ],
+            callbackUrls: ['https://redirect.com'],
+            logoutUrls: ['https://logout.com'],
+          },
+        },
+      });
+      const template = Template.fromStack(stack);
+      template.hasResourceProperties('AWS::Cognito::UserPool', {
+        UsernameAttributes: ['email'],
+        AutoVerifiedAttributes: ['email'],
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        AttributeMapping: {
+          email: 'email',
+        },
+        IdpIdentifiers: mockIdentifiers,
+        ProviderDetails: {
+          attributes_request_method: 'GET',
+          attributes_url: userInfoURL,
+          authorize_scopes: mockScopes.join(' '),
+          authorize_url: authorizationURL,
+          client_id: oidcClientId,
+          client_secret: oidcClientSecret,
+          jwks_uri: jwksURI,
+          oidc_issuer: oidcIssuerUrl,
+          token_url: tokensURL,
+        },
+        ProviderName: oidcProviderName,
+        ProviderType: 'OIDC',
+      });
+      template.hasResourceProperties('AWS::Cognito::IdentityPool', {
+        OpenIdConnectProviderARNs: [
+          Match.objectEquals({
+            'Fn::Join': [
+              '',
+              [
+                'arn:aws:iam:',
+                { Ref: 'AWS::Region' },
+                ':',
+                { Ref: 'AWS::AccountId' },
+                ':oidc-provider/cognito-idp.',
+                { Ref: 'AWS::Region' },
+                '.amazonaws.com/',
+                { Ref: 'testMyOidcProviderOidcIDP837BDEAD' },
               ],
             ],
           }),
@@ -1281,12 +1418,14 @@ void describe('Auth construct', () => {
         loginWith: {
           phone: true,
           externalProviders: {
-            oidc: {
-              clientId: oidcClientId,
-              clientSecret: oidcClientSecret,
-              issuerUrl: oidcIssuerUrl,
-              name: oidcProviderName,
-            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+              },
+            ],
             callbackUrls: ['https://redirect.com'],
             logoutUrls: ['https://logout.com'],
           },
@@ -1314,7 +1453,81 @@ void describe('Auth construct', () => {
                 ':oidc-provider/cognito-idp.',
                 { Ref: 'AWS::Region' },
                 '.amazonaws.com/',
-                { Ref: 'testOidcIDP12B3582F' },
+                { Ref: 'testMyOidcProviderOidcIDP837BDEAD' },
+              ],
+            ],
+          }),
+        ],
+      });
+    });
+    void it('supports multiple oidc providers', () => {
+      const app = new App();
+      const stack = new Stack(app);
+      new AmplifyAuth(stack, 'test', {
+        loginWith: {
+          email: true,
+          externalProviders: {
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+              },
+              {
+                clientId: oidcClientId2,
+                clientSecret: oidcClientSecret2,
+                issuerUrl: oidcIssuerUrl2,
+                name: oidcProviderName2,
+              },
+            ],
+            callbackUrls: ['https://redirect.com'],
+            logoutUrls: ['https://logout.com'],
+          },
+        },
+      });
+      const template = Template.fromStack(stack);
+      template.hasResourceProperties('AWS::Cognito::UserPool', {
+        UsernameAttributes: ['email'],
+        AutoVerifiedAttributes: ['email'],
+      });
+      template.hasResourceProperties(
+        'AWS::Cognito::UserPoolIdentityProvider',
+        ExpectedOidcIDPProperties
+      );
+      template.hasResourceProperties(
+        'AWS::Cognito::UserPoolIdentityProvider',
+        ExpectedOidcIDPProperties2
+      );
+      template.hasResourceProperties('AWS::Cognito::IdentityPool', {
+        OpenIdConnectProviderARNs: [
+          Match.objectEquals({
+            'Fn::Join': [
+              '',
+              [
+                'arn:aws:iam:',
+                { Ref: 'AWS::Region' },
+                ':',
+                { Ref: 'AWS::AccountId' },
+                ':oidc-provider/cognito-idp.',
+                { Ref: 'AWS::Region' },
+                '.amazonaws.com/',
+                { Ref: 'testMyOidcProviderOidcIDP837BDEAD' },
+              ],
+            ],
+          }),
+          Match.objectEquals({
+            'Fn::Join': [
+              '',
+              [
+                'arn:aws:iam:',
+                { Ref: 'AWS::Region' },
+                ':',
+                { Ref: 'AWS::AccountId' },
+                ':oidc-provider/cognito-idp.',
+                { Ref: 'AWS::Region' },
+                '.amazonaws.com/',
+                { Ref: 'testMyOidcProvider2OidcIDP43D7B07B' },
               ],
             ],
           }),
@@ -1600,12 +1813,14 @@ void describe('Auth construct', () => {
               clientId: amazonClientId,
               clientSecret: amazonClientSecret,
             },
-            oidc: {
-              clientId: oidcClientId,
-              clientSecret: oidcClientSecret,
-              issuerUrl: oidcIssuerUrl,
-              name: oidcProviderName,
-            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+              },
+            ],
             saml: {
               name: samlProviderName,
               metadata: {
@@ -1682,12 +1897,14 @@ void describe('Auth construct', () => {
               clientId: amazonClientId,
               clientSecret: amazonClientSecret,
             },
-            oidc: {
-              clientId: oidcClientId,
-              clientSecret: oidcClientSecret,
-              issuerUrl: oidcIssuerUrl,
-              name: oidcProviderName,
-            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+              },
+            ],
             callbackUrls: ['https://redirect.com'],
             logoutUrls: ['https://logout.com'],
           },
@@ -1761,12 +1978,14 @@ void describe('Auth construct', () => {
               clientId: amazonClientId,
               clientSecret: amazonClientSecret,
             },
-            oidc: {
-              clientId: oidcClientId,
-              clientSecret: oidcClientSecret,
-              issuerUrl: oidcIssuerUrl,
-              name: oidcProviderName,
-            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+              },
+            ],
             callbackUrls: ['https://redirect.com'],
             logoutUrls: ['https://logout.com'],
           },
@@ -1828,6 +2047,260 @@ void describe('Auth construct', () => {
         );
       });
     });
+
+    void it('automatically maps email attributes for external providers and keeps existing configuration', () => {
+      const app = new App();
+      const stack = new Stack(app);
+      new AmplifyAuth(stack, 'test', {
+        loginWith: {
+          email: true,
+          externalProviders: {
+            google: {
+              clientId: googleClientId,
+              clientSecret: SecretValue.unsafePlainText(googleClientSecret),
+              attributeMapping: {
+                fullname: ProviderAttribute.GOOGLE_NAME,
+              },
+            },
+            facebook: {
+              clientId: facebookClientId,
+              clientSecret: facebookClientSecret,
+              attributeMapping: {
+                fullname: ProviderAttribute.FACEBOOK_NAME,
+              },
+            },
+            signInWithApple: {
+              clientId: appleClientId,
+              keyId: appleKeyId,
+              privateKey: applePrivateKey,
+              teamId: appleTeamId,
+              attributeMapping: {
+                fullname: ProviderAttribute.APPLE_NAME,
+              },
+            },
+            loginWithAmazon: {
+              clientId: amazonClientId,
+              clientSecret: amazonClientSecret,
+              attributeMapping: {
+                fullname: ProviderAttribute.AMAZON_NAME,
+              },
+            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+                attributeMapping: {
+                  fullname: {
+                    attributeName: 'name',
+                  },
+                },
+              },
+            ],
+            callbackUrls: ['https://redirect.com'],
+            logoutUrls: ['https://logout.com'],
+          },
+        },
+      });
+      const template = Template.fromStack(stack);
+      template.hasResourceProperties('AWS::Cognito::UserPool', {
+        UsernameAttributes: ['email'],
+        AutoVerifiedAttributes: ['email'],
+      });
+      const expectedAutoMappedAttributes = {
+        AttributeMapping: {
+          // 'email' is a standardized claim for oauth and oidc IDPS
+          // so we can map it to cognito's 'email' claim
+          email: 'email',
+        },
+      };
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedAmazonIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.AMAZON_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedAppleIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.APPLE_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedFacebookIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.FACEBOOK_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedGoogleIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.GOOGLE_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedOidcIDPProperties,
+        AttributeMapping: {
+          ...expectedAutoMappedAttributes.AttributeMapping,
+          name: 'name',
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::IdentityPool', {
+        SupportedLoginProviders: {
+          'www.amazon.com': amazonClientId,
+          'accounts.google.com': googleClientId,
+          'appleid.apple.com': appleClientId,
+          'graph.facebook.com': facebookClientId,
+        },
+      });
+    });
+
+    void it('should not override email attribute mapping if customer providers their own mapping', () => {
+      const app = new App();
+      const stack = new Stack(app);
+      const customEmailMapping = 'customMapping';
+      new AmplifyAuth(stack, 'test', {
+        loginWith: {
+          email: true,
+          externalProviders: {
+            google: {
+              clientId: googleClientId,
+              clientSecret: SecretValue.unsafePlainText(googleClientSecret),
+              attributeMapping: {
+                email: {
+                  attributeName: customEmailMapping,
+                },
+                fullname: ProviderAttribute.GOOGLE_NAME,
+              },
+            },
+            facebook: {
+              clientId: facebookClientId,
+              clientSecret: facebookClientSecret,
+              attributeMapping: {
+                email: {
+                  attributeName: customEmailMapping,
+                },
+                fullname: ProviderAttribute.FACEBOOK_NAME,
+              },
+            },
+            signInWithApple: {
+              clientId: appleClientId,
+              keyId: appleKeyId,
+              privateKey: applePrivateKey,
+              teamId: appleTeamId,
+              attributeMapping: {
+                email: {
+                  attributeName: customEmailMapping,
+                },
+                fullname: ProviderAttribute.APPLE_NAME,
+              },
+            },
+            loginWithAmazon: {
+              clientId: amazonClientId,
+              clientSecret: amazonClientSecret,
+              attributeMapping: {
+                email: {
+                  attributeName: customEmailMapping,
+                },
+                fullname: ProviderAttribute.AMAZON_NAME,
+              },
+            },
+            oidc: [
+              {
+                clientId: oidcClientId,
+                clientSecret: oidcClientSecret,
+                issuerUrl: oidcIssuerUrl,
+                name: oidcProviderName,
+                attributeMapping: {
+                  email: {
+                    attributeName: customEmailMapping,
+                  },
+                  fullname: {
+                    attributeName: 'name',
+                  },
+                },
+              },
+            ],
+            callbackUrls: ['https://redirect.com'],
+            logoutUrls: ['https://logout.com'],
+          },
+        },
+      });
+      const template = Template.fromStack(stack);
+      template.hasResourceProperties('AWS::Cognito::UserPool', {
+        UsernameAttributes: ['email'],
+        AutoVerifiedAttributes: ['email'],
+      });
+      const expectedAutoMappedAttributes = {
+        AttributeMapping: {
+          email: customEmailMapping,
+        },
+      };
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedAmazonIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.AMAZON_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedAppleIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.APPLE_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedFacebookIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.FACEBOOK_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedGoogleIDPProperties,
+        ...{
+          AttributeMapping: {
+            ...expectedAutoMappedAttributes.AttributeMapping,
+            name: ProviderAttribute.GOOGLE_NAME.attributeName,
+          },
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolIdentityProvider', {
+        ...ExpectedOidcIDPProperties,
+        AttributeMapping: {
+          ...expectedAutoMappedAttributes.AttributeMapping,
+          name: 'name',
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::IdentityPool', {
+        SupportedLoginProviders: {
+          'www.amazon.com': amazonClientId,
+          'accounts.google.com': googleClientId,
+          'appleid.apple.com': appleClientId,
+          'graph.facebook.com': facebookClientId,
+        },
+      });
+    });
   });
 
   void it('sets resource names based on id and name property', () => {
@@ -1860,12 +2333,14 @@ void describe('Auth construct', () => {
             clientId: amazonClientId,
             clientSecret: amazonClientSecret,
           },
-          oidc: {
-            clientId: oidcClientId,
-            clientSecret: oidcClientSecret,
-            issuerUrl: oidcIssuerUrl,
-            name: oidcProviderName,
-          },
+          oidc: [
+            {
+              clientId: oidcClientId,
+              clientSecret: oidcClientSecret,
+              issuerUrl: oidcIssuerUrl,
+              name: oidcProviderName,
+            },
+          ],
           saml: {
             name: samlProviderName,
             metadata: {