+++ scripts/check_dependencies.ts
@@ -4,9 +4,15 @@ import { DependenciesValidator } from './components/dependencies_validator.js';
 await new DependenciesValidator(
   await glob('packages/*'),
   {
-    'aws-amplify': { denyAll: true },
-    '@aws-amplify/datastore': { denyAll: true },
-    '@aws-amplify/core': { denyAll: true },
+    'aws-amplify': {
+      allowList: ['@aws-amplify/integration-tests'],
+    },
+    '@aws-amplify/datastore': {
+      allowList: ['@aws-amplify/integration-tests'],
+    },
+    '@aws-amplify/core': {
+      allowList: ['@aws-amplify/integration-tests'],
+    },
     '@aws-amplify/cli-core': {
       allowList: [
         '@aws-amplify/backend-cli',