+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @aws-amplify/backend-auth
 
+## 0.4.8-beta.1
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.3.1-beta.0
+- @aws-amplify/auth-construct-alpha@0.5.7-beta.1
+
+## 0.4.8-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3cb2a527f]
+  - @aws-amplify/auth-construct-alpha@0.5.7-beta.0
+
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [29fb32bcf]
+  - @aws-amplify/auth-construct-alpha@0.5.6
+
 ## 0.4.6
 
 ### Patch Changes