+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/backend-function
 
+## 0.7.2-beta.0
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.3.1-beta.0
+
+## 0.7.1
+
+### Patch Changes
+
+- adb50ecb6: Fix cjs/top-level-await/esbuild interoperability issue when bundling lambda functions
+
 ## 0.7.0
 
 ### Minor Changes