+++ packages/backend-secret/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/backend-secret
 
+## 0.4.5-beta.0
+
+### Patch Changes
+
+- Updated dependencies [74cbda0]
+  - @aws-amplify/platform-core@0.5.0-beta.0
+
 ## 0.4.4
 
 ### Patch Changes